skin_lesion_classification/requirements.txt
python==3.8.19
torch==2.2.2+cu118
torchvision==0.17.2+cu118
numpy==1.24.1
seaborn==0.13.2
Pillow==10.2.0
pytorch-lightning==2.2.1
matplotlib==3.7.5
efficientnet-pytorch==0.7.1
pandas==2.0.3
scikit-learn==1.3.2